+++ layrz/sdk/lcl/__init__.py
@@ -1,2 +1,2 @@
-""" Init file """
+""" Layrz Compute Language (LCL) SDK """
 from .core import LclCore