+++ layrz/sdk/lcl/core.py
@@ -5,32 +5,59 @@ INVALID_NUMBER_OF_PARAMS = 'Invalid number of arguments - Expected {expected} - 
 DIFFERENT_TYPES_RANGES = 'Invalid data range, value: {arg1} - Minimum: {arg2} - Maximum: {arg3}'
 DIFFERENT_TYPES = 'Invalid data types - arg1: {arg1} - arg2: {arg2}'
 INVALID_ARGUMENTS = 'Invalid arguments - {e}'
-TIMEOUT_REACHED = 'Timeout reached - Maximum execution time: {timeout}s'
+
 
 class LclCore:
-  """
-  Layrz Compute Language
-  """
+  """ Layrz Compute Language SDK """
 
-  def __init__(self, script='', sensors={}, previous_sensors={}, payload={}, asset_constants={}, custom_fields={}):
-    """ Constructor """
-    self.__sensors = sensors
-    self.__previous_sensors = previous_sensors
-    self.__payload = payload
-    self.__asset_constants = asset_constants
-    self.__custom_fields = custom_fields
-    self.__script = script
+  def __init__(
+    self,
+    script: str = '',
+    sensors: dict = {},
+    previous_sensors: dict = {},
+    payload: dict = {},
+    asset_constants: dict = {},
+    custom_fields: dict = {},
+  ):
+    """
+    Creates a new instance of LclCore
+    ---
+    Arguments
+      - script: Is the LCL script to be executed
+      - sensors: Is the current sensor data
+      - previous_sensors: Is the previous sensor data
+      - payload: Is the payload data
+      - asset_constants: Is the asset constants data
+      - custom_fields: Is the custom fields data
+    """
+    self._sensors = sensors
+    self._previous_sensors = previous_sensors
+    self._payload = payload
+    self._asset_constants = asset_constants
+    self._custom_fields = custom_fields
+    self._script = script
 
-  def perform(self, ignore_signal=False):
-    """ Perform script using Layrz Compute Language """
+  def perform(
+    self,
+    additional_globals: dict = {},
+    additional_locals: dict = {},
+  ):
+    """
+    Perform script using Layrz Compute Language
+    ---
+    Arguments
+      - additional_globals: Additional global variables and functions
+      - additional_locals: Additional local variables
+    """
     try:
       local_variables = {
-        'payload': self.__payload,
-        'sensors': self.__sensors,
-        'custom_fields': self.__custom_fields,
-        'previous_sensors': self.__previous_sensors,
-        'asset_constants': self.__asset_constants,
+        'payload': self._payload,
+        'sensors': self._sensors,
+        'custom_fields': self._custom_fields,
+        'previous_sensors': self._previous_sensors,
+        'asset_constants': self._asset_constants,
       }
+      local_variables.update(additional_locals)
 
       ## Define global variables and functions
       global_functions = {
@@ -84,108 +111,66 @@ class LclCore:
         'SUBSTRING': self.SUBSTRING,
         'UNIX_TO_STR': self.UNIX_TO_STR,
       }
-
-      if not ignore_signal:
-        import signal
-        signal.signal(signal.SIGALRM, self.__raise_timeout)
-        signal.alarm(self.timeout)
+      global_functions.update(additional_globals)
 
       import json
-      result = json.dumps(eval(self.__script, global_functions, local_variables))
-
-      if not ignore_signal:
-        signal.signal(signal.SIGALRM, signal.SIG_IGN)
+      result = json.dumps(eval(self._script, global_functions, local_variables))
 
       return result
-    except TimeoutError:
-      return json.dumps(TIMEOUT_REACHED.format(timeout=self.timeout))
     except Exception as err:
       import json
       return json.dumps(INVALID_ARGUMENTS.format(e=err))
 
-  @property
-  def timeout(self):
-    """ Timeout constant """
-    return 10 # seconds
-
-  def __raise_timeout(self, signum, frame):
-    """ Raise timeout """
-    raise TimeoutError
-
   def GET_PARAM(self, *args):
     """ GET_PARAM Function """
     if len(args) > 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
 
     if len(args) > 1:
-      return self.__payload.get(args[0], args[1])
-    return self.__payload.get(args[0], None)
+      return self._payload.get(args[0], args[1])
+    return self._payload.get(args[0], None)
 
   def GET_DISTANCE_TRAVELED(self, *args):
     """ GET_DISTANCE_TRAVELED Function """
     if len(args) > 0:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=0,
-        received=len(args)
-      )
-    return self.__asset_constants.get('distanceTraveled', 0)
+      return INVALID_NUMBER_OF_PARAMS.format(expected=0, received=len(args))
+    return self._asset_constants.get('distanceTraveled', 0)
 
   def GET_PREVIOUS_SENSOR(self, *args):
     """ GET_PREVIOUS_SENSOR Function """
     if len(args) < 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     if len(args) > 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
 
     if len(args) > 1:
-      return self.__previous_sensors.get(args[0], args[1])
-    return self.__previous_sensors.get(args[0], None)
+      return self._previous_sensors.get(args[0], args[1])
+    return self._previous_sensors.get(args[0], None)
 
   def GET_SENSOR(self, *args):
     """ GET_SENSOR Function """
     if len(args) < 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     if len(args) > 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
 
     if len(args) > 1:
-      return self.__sensors.get(args[0], args[1])
-    return self.__sensors.get(args[0], None)
+      return self._sensors.get(args[0], args[1])
+    return self._sensors.get(args[0], None)
 
   def CONSTANT(self, *args):
     """ CONSTANT Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     return args[0]
 
   def GET_CUSTOM_FIELD(self, *args):
     """ GET_CUSTOM_FIELD Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
-    return self.__custom_fields.get(args[0], '')
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
+    return self._custom_fields.get(args[0], '')
 
   def COMPARE(self, *args):
     """ COMPARE Function """
@@ -301,10 +286,7 @@ class LclCore:
   def TO_BOOL(self, *args):
     """ TO_BOOL Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     return bool(args[0])
@@ -312,10 +294,7 @@ class LclCore:
   def TO_STR(self, *args):
     """ TO_STR Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     if args[0] is None:
       return None
@@ -324,10 +303,7 @@ class LclCore:
   def TO_INT(self, *args):
     """ TO_INT Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     return int(args[0])
@@ -335,10 +311,7 @@ class LclCore:
   def CEIL(self, *args):
     """ CEIL Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     import math
@@ -347,10 +320,7 @@ class LclCore:
   def FLOOR(self, *args):
     """ FLOOR Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     import math
@@ -359,10 +329,7 @@ class LclCore:
   def ROUND(self, *args):
     """ ROUND Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     return round(args[0])
@@ -370,10 +337,7 @@ class LclCore:
   def SQRT(self, *args):
     """ SQRT Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     import math
@@ -389,15 +353,9 @@ class LclCore:
   def RANDOM(self, *args):
     """ RANDOM Function """
     if len(args) > 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
     if len(args) < 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
 
     import random
     return random.random() * (float(args[1]) - float(args[0])) + float(args[0])
@@ -405,10 +363,7 @@ class LclCore:
   def RANDOM_INT(self, *args):
     """ RANDOM_INT Function """
     if len(args) != 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
 
     import random
     return random.randint(int(args[0]), int(args[1]))
@@ -416,125 +371,77 @@ class LclCore:
   def GREATER_THAN_OR_EQUALS_TO(self, *args):
     """ GREATER_THAN_OR_EQUALS_TO Function """
     if len(args) > 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if len(args) < 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     if args[0] is None or args[1] is None:
       return None
 
     if type(args[0]) != type(args[1]):
-      return DIFFERENT_TYPES.format(
-        arg1=type(args[0]),
-        arg2=type(args[1])
-      )
+      return DIFFERENT_TYPES.format(arg1=type(args[0]), arg2=type(args[1]))
     return args[0] >= args[1]
 
   def GREATER_THAN(self, *args):
     """ GREATER_THAN Function """
     if len(args) > 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if len(args) < 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     if args[0] is None or args[1] is None:
       return None
 
     if type(args[0]) != type(args[1]):
-      return DIFFERENT_TYPES.format(
-        arg1=type(args[0]),
-        arg2=type(args[1])
-      )
+      return DIFFERENT_TYPES.format(arg1=type(args[0]), arg2=type(args[1]))
     return args[0] > args[1]
 
   def LESS_THAN_OR_EQUALS_TO(self, *args):
     """ LESS_THAN_OR_EQUALS_TO Function """
     if len(args) > 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if len(args) < 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     if args[0] is None or args[1] is None:
       return None
 
     if type(args[0]) != type(args[1]):
-      return DIFFERENT_TYPES.format(
-        arg1=type(args[0]),
-        arg2=type(args[1])
-      )
+      return DIFFERENT_TYPES.format(arg1=type(args[0]), arg2=type(args[1]))
     return args[0] <= args[1]
 
   def LESS_THAN(self, *args):
     """ LESS_THAN Function """
     if len(args) > 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if len(args) < 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     if args[0] is None or args[1] is None:
       return None
 
     if type(args[0]) != type(args[1]):
-      return DIFFERENT_TYPES.format(
-        arg1=type(args[0]),
-        arg2=type(args[1])
-      )
+      return DIFFERENT_TYPES.format(arg1=type(args[0]), arg2=type(args[1]))
     return args[0] < args[1]
 
   def DIFFERENT(self, *args):
     """ DIFFERENT Function """
     if len(args) > 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if len(args) < 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     if args[0] is None or args[1] is None:
       return None
 
     if type(args[0]) != type(args[1]):
-      return DIFFERENT_TYPES.format(
-        arg1=type(args[0]),
-        arg2=type(args[1])
-      )
+      return DIFFERENT_TYPES.format(arg1=type(args[0]), arg2=type(args[1]))
     return args[0] != args[1]
 
   def HEX_TO_STR(self, *args):
     """ HEX_TO_STR Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     if args[0] is None:
       return None
@@ -544,10 +451,7 @@ class LclCore:
   def STR_TO_HEX(self, *args):
     """ STR_TO_HEX Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     return str(args[0]).encode('ASCII').hex()
@@ -555,10 +459,7 @@ class LclCore:
   def HEX_TO_INT(self, *args):
     """ HEX_TO_INT Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     return int(int(args[0], 16))
@@ -566,10 +467,7 @@ class LclCore:
   def INT_TO_HEX(self, *args):
     """ INT_TO_HEX Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     return hex(int(args[0]))[2:]
@@ -577,10 +475,7 @@ class LclCore:
   def TO_FLOAT(self, *args):
     """ TO_FLOAT Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
     if args[0] is None:
       return None
     return float(args[0])
@@ -588,77 +483,51 @@ class LclCore:
   def IS_PARAMETER_PRESENT(self, *args):
     """ IS_PARAMETER_PRESENT Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
-    return args[0] in self.__payload
+    return args[0] in self._payload
 
   def IS_SENSOR_PRESENT(self, *args):
     """ IS_SENSOR_PRESENT Function """
     if len(args) > 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
-    return args[0] in self.__sensors
+    return args[0] in self._sensors
 
   def INSIDE_RANGE(self, *args):
     """ INSIDE_RANGE Function """
     if len(args) != 3:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=3,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=3, received=len(args))
 
     if args[0] is None or args[1] is None or args[2] is None:
       return None
     if type(args[0]) != type(args[1]) or type(args[0]) != type(args[2]):
-      return DIFFERENT_TYPES_RANGES.format(
-        arg1=type(args[0]),
-        arg2=type(args[1]),
-        arg3=type(args[2])
-      )
+      return DIFFERENT_TYPES_RANGES.format(arg1=type(args[0]), arg2=type(args[1]), arg3=type(args[2]))
 
     return args[1] <= args[0] <= args[2]
 
   def OUTSIDE_RANGE(self, *args):
     """ OUTSIDE_RANGE Function """
     if len(args) != 3:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=3,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=3, received=len(args))
 
     if args[0] is None or args[1] is None or args[2] is None:
       return None
     if type(args[0]) != type(args[1]) or type(args[0]) != type(args[2]):
-      return DIFFERENT_TYPES_RANGES.format(
-        arg1=type(args[0]),
-        arg2=type(args[1]),
-        arg3=type(args[2])
-      )
+      return DIFFERENT_TYPES_RANGES.format(arg1=type(args[0]), arg2=type(args[1]), arg3=type(args[2]))
 
     return not (args[1] <= args[0] <= args[2])
 
   def GET_TIME_DIFFERENCE(self, *args):
     """ GET_TIME_DIFFERENCE Function """
     if len(args) > 0:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=0,
-        received=len(args)
-      )
-    return self.__asset_constants.get('timeElapsed', 0)
+      return INVALID_NUMBER_OF_PARAMS.format(expected=0, received=len(args))
+    return self._asset_constants.get('timeElapsed', 0)
 
   def IF(self, *args):
     """ IF Function """
     if len(args) != 3:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=3,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=3, received=len(args))
 
     return args[1] if args[0] else args[2]
 
@@ -672,10 +541,7 @@ class LclCore:
   def REGEX(self, *args):
     """ REGEX Function """
     if len(args) != 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
 
     if args[0] is None or args[1] is None:
       return None
@@ -690,62 +556,44 @@ class LclCore:
   def IS_NONE(self, *args):
     """ IS_NONE Function """
     if len(args) != 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     return args[0] is None
 
   def NOT(self, *args):
     """ NOT Function """
     if len(args) != 1:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=1,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=1, received=len(args))
 
     return not args[0]
 
   def CONTAINS(self, *args):
     """ CONTAINS function """
     if len(args) != 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
 
     return str(args[0]) in str(args[1])
 
   def STARTS_WITH(self, *args):
     """ STARTS_WITH function """
     if len(args) != 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
 
     return str(args[1]).startswith(str(args[0]))
 
   def ENDS_WITH(self, *args):
     """ ENDS_WITH function """
     if len(args) != 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
 
     return str(args[1]).endswith(str(args[0]))
 
   def PRIMARY_DEVICE(self, *args):
     """ PRIMARY_DEVICE function """
     if len(args) > 0:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=0,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=0, received=len(args))
 
-    return self.__asset_constants.get('primaryDevice', None)
+    return self._asset_constants.get('primaryDevice', None)
 
   def SUBSTRING(self, *args):
     """ Get a substring from string (args[0]) """
@@ -754,40 +602,28 @@ class LclCore:
         expected=2,
         received=len(args),
       )
-      
+
     if len(args) > 3:
       return INVALID_NUMBER_OF_PARAMS.format(
         expected=3,
         received=len(args),
       )
-      
+
     if args[0] is None or not isinstance(args[0], str):
-      return DIFFERENT_TYPES.format(
-        arg1='str',
-        arg2=type(args[0])
-      )
-      
+      return DIFFERENT_TYPES.format(arg1='str', arg2=type(args[0]))
+
     if args[1] is None or not isinstance(args[1], int):
-      return DIFFERENT_TYPES.format(
-        arg1='int',
-        arg2=type(args[1])
-      )
-    
+      return DIFFERENT_TYPES.format(arg1='int', arg2=type(args[1]))
+
     if len(args) == 3:
       if args[2] is None or not isinstance(args[2], int):
-        return DIFFERENT_TYPES.format(
-          arg1='str',
-          arg2=type(args[2])
-        )
+        return DIFFERENT_TYPES.format(arg1='str', arg2=type(args[2]))
       return args[0][args[1]:args[2]]
     return args[0][args[1]:]
 
   def UNIX_TO_STR(self, *args):
     """ Convert UNIX timestamp date (args[0]) to format (args[1]) string """
     if len(args) != 2:
-      return INVALID_NUMBER_OF_PARAMS.format(
-        expected=2,
-        received=len(args)
-      )
+      return INVALID_NUMBER_OF_PARAMS.format(expected=2, received=len(args))
     from datetime import datetime
     return datetime.fromtimestamp(int(args[0])).strftime(args[1])